hitman-api/pyproject.toml
[tool.poetry]
name = "hitman"
version = "0.1.0"
description = ""
authors = ["William De la Cruz <macwilliamdlc@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.1.1"
uwsgi = "^2.0.19"
django-environ = "^0.4.5"
psycopg2-binary = "^2.8.6"
black = "^20.8b1"
django-username-email = "^2.4.0"
djangorestframework = "^3.11.1"
django-cors-headers = "^3.5.0"
factory_boy = "^3.0.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
